refactor(position): Position and Direction to position package
