+++ src/main/java/chess2/domain2/board2/piece2/NonPawn.java
@@ -1,6 +1,6 @@
 package chess2.domain2.board2.piece2;
 
-import chess2.domain2.board2.Position;
+import chess2.domain2.board2.position.Position;
 
 public final class NonPawn extends Piece {
