+++ src/main/java/chess2/domain2/game2/GameResult.java
@@ -7,7 +7,7 @@ import static chess2.domain2.board2.piece2.PieceType.PAWN;
 import static chess2.util2.PositionConverterUtil.FILES_TOTAL_SIZE;
 import static chess2.util2.PositionConverterUtil.RANKS_TOTAL_SIZE;
 
-import chess2.domain2.board2.Position;
+import chess2.domain2.board2.position.Position;
 import chess2.domain2.board2.piece2.Color;
 import chess2.domain2.board2.piece2.Piece;
 import chess2.util2.PieceScoreUtil;
