+++ src/main/java/chess2/domain2/game2/Running.java
@@ -3,7 +3,7 @@ package chess2.domain2.game2;
 import static chess2.domain2.board2.piece2.PieceType.KING;
 
 import chess2.domain2.board2.Board;
-import chess2.domain2.board2.Position;
+import chess2.domain2.board2.position.Position;
 import chess2.domain2.board2.piece2.Color;
 import chess2.dto2.MoveCommandDto;
