+++ src/main/java/chess2/strategy/RouteChecker.java
@@ -1,6 +1,6 @@
 package chess2.strategy;
 
-import chess2.domain2.board2.Position;
+import chess2.domain2.board2.position.Position;
 
 @FunctionalInterface
 public interface RouteChecker {
