+++ src/main/java/chess2/util2/PositionUtil.java
@@ -1,6 +1,6 @@
 package chess2.util2;
 
-import chess2.domain2.board2.Position;
+import chess2.domain2.board2.position.Position;
 import java.util.ArrayList;
 import java.util.List;
